# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: AI Software Planning Agent API
  description: |
    AI Software Planning Agent 帮助用户将模糊的软件想法逐步收敛为可执行的软件项目方案。

    ## 核心能力
    - 多轮需求沟通
    - 多模型并行分析 (GLM-5.1 / DeepSeek-V4-Pro / MiniMax-M2.5)
    - 需求融合
    - 可行性分析
    - 风险分析
    - MVP 收缩
    - 平台推荐
    - 项目规划生成 (11 类产物)

    ## 认证
    所有 API (除 `/health` 和 `/models`) 需携带 Bearer Token:
    ```
    Authorization: Bearer <api_key>
    ```

    ## 通用规范
    - 时间格式: ISO 8601 UTC (`2024-06-24T10:00:00Z`)
    - ID 格式: UUID v4
    - 分页: `offset` + `limit` (默认 20, 最大 100)
    - 错误格式: `{ error: { code, message, details } }`
  version: "1.0.0"
  contact:
    name: AI Planning Agent Team
servers:
  - url: http://localhost:3001/api/v1
    description: 本地开发环境
  - url: https://api.ai-planning-agent.com/api/v1
    description: 生产环境
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API Key 认证。在环境变量 `API_KEY` 为空时跳过认证。

  # ─── Schemas ───
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 错误码
              enum:
                - INVALID_INPUT
                - PROJECT_NOT_FOUND
                - CONVERSATION_NOT_FOUND
                - ARTIFACT_NOT_FOUND
                - WORKFLOW_NOT_FOUND
                - EXECUTION_NOT_FOUND
                - INVALID_STAGE_TRANSITION
                - WORKFLOW_ALREADY_RUNNING
                - WORKFLOW_NOT_RUNNING
                - WORKFLOW_STAGE_NOT_CLARIFICATION
                - LLM_TIMEOUT
                - LLM_ERROR
                - ALL_MODELS_FAILED
                - COST_LIMIT_EXCEEDED
                - RATE_LIMITED
                - EXPORT_NOT_READY
                - EXPORT_FAILED
                - INTERNAL_ERROR
            message:
              type: string
              description: 人类可读的错误信息
            details:
              type: object
              additionalProperties: true
              description: 错误详情 (可选)

    # ─── 枚举 ───
    ProjectStatus:
      type: string
      enum: [active, completed, failed]
    WorkflowStage:
      type: string
      enum:
        - init
        - requirement_analysis
        - requirement_clarification
        - multi_model_analysis
        - requirement_synthesis
        - feasibility_analysis
        - risk_analysis
        - mvp_compression
        - platform_recommendation
        - planning_generation
        - completed
        - failed
    StageStatus:
      type: string
      enum: [pending, running, completed, failed, skipped]
    ExecutionStatus:
      type: string
      enum: [success, failed, timeout, cancelled]
    CallStatus:
      type: string
      enum: [success, failed, timeout, rate_limited]
...
    WorkflowProgress:
      type: object
      required: [completed_stages, total_stages, percentage]
      properties:
        completed_stages:
          type: integer
          minimum: 0
          maximum: 9
        total_stages:
          type: integer
          const: 9
        percentage:
          type: number
          minimum: 0
          maximum: 100
    ClarificationQuestion:
      type: object
      required: [question, context, category]
      properties:
        question:
          type: string
          description: 澄清问题
        context:
          type: string
          description: 为什么需要澄清
        category:
          type: string
          enum: [user, scope, tech, business, risk]
          description: 问题类别
    ModelStatus:
      type: object
      properties:
        deepseek:
          type: string
          enum: [running, completed, failed, pending]
        glm:
          type: string
          enum: [running, completed, failed, pending]
        minimax:
          type: string
          enum: [running, completed, failed, pending]
    WorkflowStatusResponse:
      type: object
      required: [project_id, conversation_id, status, current_stage, stage_display_name, progress, updated_at]
      properties:
        project_id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
          nullable: true
          description: 当前工作流检查点会话 ID；尚未开始时为空
        status:
          $ref: "#/components/schemas/ProjectStatus"
        current_stage:
          $ref: "#/components/schemas/WorkflowStage"
        stage_display_name:
          type: string
          description: 当前阶段的人类可读名称
          example: "需求分析"
        progress:
          $ref: "#/components/schemas/WorkflowProgress"
        waiting_for:
          type: string
          nullable: true
          enum: [reply, review]
          description: reply 表示等待补充信息；review 表示可讨论或确认推进
        next_stage:
          $ref: "#/components/schemas/WorkflowStage"
          nullable: true
          description: 确认当前检查点后将进入的阶段
        clarification_questions:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ClarificationQuestion"
          description: 澄清问题 (仅在 clarification 阶段)
        model_status:
          $ref: "#/components/schemas/ModelStatus"
          nullable: true
          description: 多模型状态 (仅在 multi_model_analysis 阶段)
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
    ContinueWorkflowRequest:
      type: object
      required: [conversation_id, message]
      properties:
        conversation_id:
          type: string
          format: uuid
          description: 对话 ID
        message:
          type: string
          minLength: 1
          maxLength: 50000
          description: 用户对澄清问题的回复
    AdvanceWorkflowRequest:
      type: object
      required: [conversation_id]
      properties:
        conversation_id:
          type: string
          format: uuid
          description: 当前检查点会话 ID

    # ─── Workflow State ───
    WorkflowStateResponse:
      type: object
      required: [id, project_id, stage, status, display_name, progress, created_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        stage:
          $ref: "#/components/schemas/WorkflowStage"
        status:
          $ref: "#/components/schemas/StageStatus"
        display_name:
          type: string
          example: "需求分析"
        progress:
          $ref: "#/components/schemas/WorkflowProgress"
        data_json:
          type: object
          nullable: true
          description: 阶段输出数据
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    WorkflowStateListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStateResponse"
        total:
          type: integer

    # ─── Workflow Execution ───
    WorkflowExecutionResponse:
      type: object
      required: [id, project_id, stage, status, started_at, retry_count, created_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        stage:
          type: string
          description: 执行的阶段
        status:
          $ref: "#/components/schemas/ExecutionStatus"
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        duration_ms:
          type: integer
          nullable: true
          description: 执行耗时 (ms)
        error_message:
          type: string
          nullable: true
        retry_count:
          type: integer
          description: 阶段级重试次数
        model_call_count:
          type: integer
          nullable: true
          description: 本次执行中的模型调用次数
        created_at:
          type: string
          format: date-time
    WorkflowExecutionListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowExecutionResponse"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    WorkflowExecutionDetailResponse:
      allOf:
        - $ref: "#/components/schemas/WorkflowExecutionResponse"
        - type: object
          properties:
            model_logs:
              type: array
              items:
                $ref: "#/components/schemas/ModelExecutionLogResponse"
              description: 关联的模型调用日志

    # ─── Model Execution Log ───
    ModelExecutionLogResponse:
      type: object
      required:
        - id
        - project_id
        - stage
        - provider_name
        - model_id
        - status
        - attempt_number
        - input_tokens
        - output_tokens
        - cost_total
        - created_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        execution_id:
          type: string
          format: uuid
          nullable: true
        stage:
          type: string
        provider_name:
          type: string
          description: Provider 逻辑名称
          example: "deepseek"
        model_id:
          type: string
          description: 模型 ID
          example: "deepseek-v4-pro"
        status:
          $ref: "#/components/schemas/CallStatus"
        attempt_number:
          type: integer
          description: 第几次尝试 (1-based)
        prompt_version_id:
          type: string
          format: uuid
          nullable: true
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cost_input:
          type: number
          description: 输入成本 (元)
        cost_output:
          type: number
          description: 输出成本 (元)
        cost_total:
          type: number
          description: 总成本 (元)
        latency_ms:
          type: integer
          nullable: true
          description: 调用延迟 (ms)
        error_code:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ModelExecutionLogListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModelExecutionLogResponse"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer

    # ─── Artifact ───
...
paths:
...
  /projects/{project_id}/run:
    post:
      tags: [Workflow]
      summary: 启动工作流
      description: |
        启动项目的完整工作流 Pipeline。
        工作流按以下顺序自动执行:
        INIT → REQUIREMENT_ANALYSIS → REQUIREMENT_CLARIFICATION (暂停)
        → MULTI_MODEL_ANALYSIS → REQUIREMENT_SYNTHESIS
        → FEASIBILITY_ANALYSIS → RISK_ANALYSIS → MVP_COMPRESSION
        → PLATFORM_RECOMMENDATION → PLANNING_GENERATION → COMPLETED

        仅在 REQUIREMENT_CLARIFICATION 阶段暂停，等待用户输入。
        其他阶段自动推进。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 项目 ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation_id:
                  type: string
                  format: uuid
                  description: 关联的对话 ID (可选，不传则自动创建)
      responses:
        "202":
          description: 工作流已启动
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStatusResponse"
              example:
                project_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "active"
                current_stage: "requirement_analysis"
                stage_display_name: "需求分析"
                progress:
                  completed_stages: 0
                  total_stages: 9
                  percentage: 0
                started_at: "2024-06-24T10:00:00Z"
                updated_at: "2024-06-24T10:00:00Z"
        "404":
          description: 项目不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 工作流已在运行中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "WORKFLOW_ALREADY_RUNNING"
                  message: "Workflow is already running for this project"
  /projects/{project_id}/workflow/status:
    get:
      tags: [Workflow]
      summary: 获取工作流状态
      description: |
        获取当前工作流的实时状态。
        前端应轮询此接口 (建议间隔: 2s-5s，取决于阶段)。
        clarification 阶段返回 clarification_questions，
        multi_model_analysis 阶段返回 model_status。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 工作流状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStatusResponse"
              examples:
                running:
                  summary: 执行中
                  value:
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    status: "active"
                    current_stage: "multi_model_analysis"
                    stage_display_name: "多模型分析"
                    progress:
                      completed_stages: 2
                      total_stages: 9
                      percentage: 22.2
                    model_status:
                      deepseek: "running"
                      glm: "running"
                      minimax: "running"
                    updated_at: "2024-06-24T10:00:30Z"
                clarification:
                  summary: 澄清阶段 (等待用户输入)
                  value:
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    status: "active"
                    current_stage: "requirement_clarification"
                    stage_display_name: "需求澄清"
                    progress:
                      completed_stages: 1
                      total_stages: 9
                      percentage: 11.1
                    clarification_questions:
                      - question: "What platform should this be built for?"
                        context: "Platform choice affects tech stack and architecture"
                        category: "tech"
                    updated_at: "2024-06-24T10:00:15Z"
                completed:
                  summary: 已完成
                  value:
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    status: "completed"
                    current_stage: "completed"
                    stage_display_name: "已完成"
                    progress:
                      completed_stages: 9
                      total_stages: 9
                      percentage: 100
                    updated_at: "2024-06-24T10:05:00Z"
                failed:
                  summary: 失败
                  value:
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    status: "failed"
                    current_stage: "failed"
                    stage_display_name: "失败"
                    progress:
                      completed_stages: 2
                      total_stages: 9
                      percentage: 22.2
                    error_message: "All models failed: deepseek(timeout), glm(timeout), minimax(timeout)"
                    updated_at: "2024-06-24T10:01:00Z"
        "404":
          description: 项目不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{project_id}/workflow/continue:
    post:
      tags: [Workflow]
      summary: 继续工作流
      description: |
        在 REQUIREMENT_CLARIFICATION 阶段，用户回复澄清问题后调用此接口继续工作流。
        Agent 会重新分析需求，决定是否需要继续澄清还是进入下一阶段。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContinueWorkflowRequest"
            example:
              conversation_id: "660e8400-e29b-41d4-a716-446655440001"
              message: "The app should support web and iOS platforms. It's for individual users only."
      responses:
        "202":
          description: 工作流继续执行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStatusResponse"
        "404":
          description: 项目或对话不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: |
            当前阶段不是 REQUIREMENT_CLARIFICATION，无法继续。
            可能原因: 工作流未启动、已不在澄清阶段、已结束。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "WORKFLOW_STAGE_NOT_CLARIFICATION"
                  message: "Cannot continue workflow from stage 'multi_model_analysis'"
                  details:
                    current_stage: "multi_model_analysis"
                    expected_stage: "requirement_clarification"
  /projects/{project_id}/workflow/discuss:
    post:
      tags: [Workflow]
      summary: 讨论当前检查点
      description: 在等待用户确认的检查点中发送一条消息，Agent 会基于当前阶段结果回复；不会自动推进工作流。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContinueWorkflowRequest"
      responses:
        "202":
          description: Agent 已回复
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStatusResponse"
  /projects/{project_id}/workflow/advance:
    post:
      tags: [Workflow]
      summary: 确认并进入下一环节
      description: 关闭当前检查点会话，并执行到下一个需要用户决策的检查点或工作流完成。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvanceWorkflowRequest"
      responses:
        "202":
          description: 工作流已推进
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStatusResponse"
  /projects/{project_id}/workflow/states:
    get:
      tags: [Workflow]
      summary: 获取工作流各阶段状态
      description: 返回项目所有阶段的状态快照列表。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 阶段状态列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStateListResponse"
        "404":
          description: 项目不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ═══════════════════════════════════════════════════════
  # Workflow Executions
  # ═══════════════════════════════════════════════════════
  /projects/{project_id}/workflow/executions:
    get:
      tags: [Workflow Executions]
      summary: 获取工作流执行历史
      description: |
        分页获取项目的所有阶段执行历史，按时间倒序。
        每次阶段执行 (包括重试) 都会生成一条记录。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: stage
          in: query
          schema:
            $ref: "#/components/schemas/WorkflowStage"
          description: 按阶段过滤
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ExecutionStatus"
          description: 按执行状态过滤
      responses:
        "200":
          description: 执行历史列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowExecutionListResponse"
              example:
                items:
                  - id: "990e8400-e29b-41d4-a716-446655440010"
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    stage: "multi_model_analysis"
                    status: "success"
                    started_at: "2024-06-24T10:00:20Z"
                    completed_at: "2024-06-24T10:00:25Z"
                    duration_ms: 5230
                    retry_count: 0
                    model_call_count: 3
                    created_at: "2024-06-24T10:00:25Z"
                  - id: "990e8400-e29b-41d4-a716-446655440009"
                    project_id: "550e8400-e29b-41d4-a716-446655440000"
                    stage: "requirement_analysis"
                    status: "failed"
                    started_at: "2024-06-24T10:00:05Z"
                    completed_at: "2024-06-24T10:00:08Z"
                    duration_ms: 3200
                    retry_count: 3
                    error_message: "LLM timeout after 3 retries"
                    model_call_count: 4
                    created_at: "2024-06-24T10:00:08Z"
                total: 15
                offset: 0
                limit: 20
        "404":
          description: 项目不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{project_id}/workflow/executions/{execution_id}:
    get:
      tags: [Workflow Executions]
      summary: 获取执行详情
      description: 获取单次阶段执行的详细信息，包含关联的模型调用日志。
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 执行详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowExecutionDetailResponse"
        "404":
          description: 项目或执行记录不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{project_id}/workflow/executions/{execution_id}/logs:
    get:
      tags: [Workflow Executions]
      summary: 获取执行关联的模型调用日志
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: 模型调用日志列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelExecutionLogListResponse"
        "404":
          description: 项目或执行记录不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ═══════════════════════════════════════════════════════
  # Conversations
  # ═══════════════════════════════════════════════════════
...